reafactor: 컴포넌트 function()으로 수정
