+++ src/pages/Community/types/WriteModal.ts
@@ -1,0 +1,12 @@
+export interface WritePostModalProps {
+  isVisible: boolean;
+  onClose: () => void;
+}
+
+export interface WritePostFormData {
+  title: string;
+  content: string;
+  file?: File;
+  isSecret: boolean;
+  allowComments: boolean;
+}
